Rename pkg/wsproxy and wsproxy.ReverseProxy
